refactor: InputHandler 패키지 수정 lotto -> lotto/util
