+++ src/main/java/lotto/domain/LottoGame.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import java.util.Optional;
-import lotto.InputHandler;
+import lotto.util.InputHandler;
 import lotto.constant.Rank;
 
 public class LottoGame {
